Do not serialize value of non-scalar FProperty
*Issue*
If the native class has a property with an array type (e.g.,
`float[]`), that will pass the current check in
`ShouldSerializePropertyValue` and cause the JSON response to have a
format that cannot be parsed in Visual Studio yet.

*Fix*
Skip properties where `ArrayDim>1`, which means they are not a
scalar value.

Note: There are planned changes to the code in VS that is reading the
JSON, but having this at the plugin level for now ensures customers can
get a workaround sooner.