[Bluprints] Change reported function references
The current logic for the blueprints commandlet is reporting functions
from the classes the blueprint _uses_ but should instead be reporting
the ones that are _implemented_ in the blueprint.

The current version is from the initial design where we would scann
_all_ the assets we could find, but since the changes in v2 to scan
based on the parent classes, this is incorrect, because it under-reports
the cases the Code Lens is supposed to show.

The current info, about the _usages_ across all assets, is now covered
by the BP References commandlet.