Removing priority mask from the test filters
Usually this priority tag is used in conjunction with other test tags (engine, product, etc.). There's no need to have this priority mask on by default since tests that have a priority tag will be marked with some other tag as well anyways. Otherwise we will be displaying prioritized tests when we turn all of the filters in Tools > Options off which looks like a bug. since. Of course, this would mean tests marked ONLY with a priority tag will not be found unless another tag is added as well (engine, product, etc.). This is not likely though, so I think this is a better approach than adding a new ‘Priority’ filter under Tools > Options. which is pretty vague.